data_processing/partnet_voxhammer/requirements.txt
numpy
Pillow
google-generativeai
